Reorder and redefine ALU related macros.
  ALUBrm had an extra comma in one of the arguments to the _O_r_x macro,
but never caused an error as it is not called from anywhere.

  The new _s32 and _u32 macros were added, to trigger constants too
large (what was actually the main initial problem in x86_64), and,
because of this new change, changed some code to cast some unsigned int
values to int, for example, argument to jit_boaddi_i.

  The ALU macros were also changed to avoid using Os_xxx_sL, and instead
check in the definition if the immediate fits in a signed 8 bytes or
not, also adjusting "locally" the opcode. This avoids several repeated
evaluations of the macro to check if the immediate fits in 8 bytes, and
also makes the code easier to follow.

  Extra evaluation of _O_r_X argument was removed as _r_X does not use
the extra argument; it is #if 0'ed, in the logic to handle x86_64 rip
addressing mode. This can/should be added back later, if really
required.

  Also in this commit, changed jit_check8() to also "recognize" %ebx
as a register that can be addressed as 8 bits in i386. It indeed is,
and actually the macro was not updated when reworking in core.h because
it was useful to test some changes, and ended up forgotten...

+++ lightning/i386/core.h
@@ -240,9 +240,9 @@ __jit_inline void
 jit_addci_ui(jit_gpr_t r0, jit_gpr_t r1, unsigned int i0)
 {
     if (r0 == r1)
-	ADDLir(i0, r0);
+	ADDLir((int)i0, r0);
     else {
-	MOVLir(i0, r0);
+	MOVLir((int)i0, r0);
 	ADDLrr(r1, r0);
     }
 }
@@ -266,9 +266,9 @@ __jit_inline void
 jit_addxi_ui(jit_gpr_t r0, jit_gpr_t r1, unsigned int i0)
 {
     if (r0 == r1)
-	ADCLir(i0, r0);
+	ADCLir((int)i0, r0);
     else {
-	MOVLir(i0, r0);
+	MOVLir((int)i0, r0);
 	ADCLrr(r1, r0);
     }
 }
@@ -291,9 +291,9 @@ __jit_inline void
 jit_subci_ui(jit_gpr_t r0, jit_gpr_t r1, unsigned int i0)
 {
     if (r0 == r1)
-	SUBLir(i0, r0);
+	SUBLir((int)i0, r0);
     else {
-	MOVLir(i0, r0);
+	MOVLir((int)i0, r0);
 	SUBLrr(r1, r0);
     }
 }
@@ -315,9 +315,9 @@ __jit_inline void
 jit_subxi_ui(jit_gpr_t r0, jit_gpr_t r1, unsigned int i0)
 {
     if (r0 == r1)
-	SBBLir(i0, r0);
+	SBBLir((int)i0, r0);
     else {
-	MOVLir(i0, r0);
+	MOVLir((int)i0, r0);
 	SBBLrr(r1, r0);
     }
 }
@@ -535,11 +535,11 @@ __jit_inline void
 jit_muli_ui_(jit_gpr_t r0, unsigned int i0)
 {
     if (r0 == _RAX) {
-	MOVLir(i0, _RDX);
+	MOVLir((int)i0, _RDX);
 	MULLr(_RDX);
     }
     else {
-	MOVLir(i0, _RAX);
+	MOVLir((int)i0, _RAX);
 	MULLr(r0);
     }
 }
@@ -1417,7 +1417,7 @@ jit_bosubr_i(jit_insn *label, jit_gpr_t r0, jit_gpr_t r1)
 __jit_inline jit_insn *
 jit_boaddi_ui(jit_insn *label, jit_gpr_t r0, unsigned int i0)
 {
-    ADDLir(i0, r0);
+    ADDLir((int)i0, r0);
     JCm(label);
     return (_jit.x.pc);
 }
@@ -1435,7 +1435,7 @@ jit_boaddr_ui(jit_insn *label, jit_gpr_t r0, jit_gpr_t r1)
 __jit_inline jit_insn *
 jit_bosubi_ui(jit_insn *label, jit_gpr_t r0, unsigned int i0)
 {
-    SUBLir(i0, r0);
+    SUBLir((int)i0, r0);
     JCm(label);
     return (_jit.x.pc);
 }
